requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
requests==2.31.0
sentry-sdk==1.31.0
python-dotenv==1.0.0
openai==0.27.9
numpy==1.23.5
pymongo==4.5.0
joblib==1.3.2
urllib3==1.26.11
gunicorn==21.2.0
gevent==23.9.1
boto3
sagemaker==2.192.1